created a helper function
this function will simplify the initializing the game board based on the different levels

also moved extra functions from the games file to the API folder
